Refactor modules into separate files
